Struct newton::body::StaticBody

pub struct StaticBody<'a> { /* fields omitted */ }

Static is the equivalent of a dynamic body with zero mass.

In newton, a static body is simply a dynamic body with no mass. This type may be removed in the future. We'll see.

Methods

impl<'a> StaticBody<'a>

pub unsafe fn from_raw(raw: *const NewtonBody, owned: bool) -> Self

pub fn create<C>(
newton: &'a Newton,
collision: &C,
matrix: [[f32; 4]; 4],
name: Option<&'static str>
) -> Self where
C: NewtonCollision,