Struct newton::collision::params::HeightFieldParams

pub struct HeightFieldParams<T> { /* fields omitted */ }

Definition of a HeightField collision

Methods

impl<T> HeightFieldParams<T>

pub fn set_grid(&mut self, grid: GridConstruction)

pub fn grid(&self) -> GridConstruction

pub fn set_scale_x(&mut self, scale: f32)

pub fn set_scale_y(&mut self, scale: f32)

pub fn set_scale_z(&mut self, scale: f32)

pub fn set_scale_xz(&mut self, x: f32, z: f32)

pub fn rows(&self) -> usize

pub fn columns(&self) -> usize

pub fn scale(&self) -> (f32, f32, f32)

The scale along (x, y, z) dimensions

pub fn elevation(&self) -> &Field<T>

pub fn elevation_mut(&mut self) -> &mut Field<T>

pub fn ids(&self) -> &Field<i8>

pub fn ids_mut(&mut self) -> &mut Field<i8>

impl<T: Num> HeightFieldParams<T>

pub fn new(rows: usize, columns: usize) -> Self

pub fn set_max_height(&mut self, max_height: f32)

Sets the maximum height, assuming that the maximum range of the elevation map is [0.0, 1.0]

pub fn max_height(&self) -> f32

Returns the max height, which is equivalent to scale_y * T::MAX_VALUE