Struct newton::collision::Compound

pub struct Compound<'a> { /* fields omitted */ }

Methods

impl<'a> Compound<'a>

pub unsafe fn from_raw(raw: *const NewtonCollision, owned: bool) -> Self

pub fn create_instance(col: &Self) -> Self

pub fn release(self)

Forgets about this body. It will be freed along the world If you want to reference the object later, you should call into_handle instead.

impl<'a> Compound<'a>

pub fn create(newton: &'a Newton) -> Self

Important traits for Handles<'a>
impl<'a> Iterator for Handles<'a> type Item = Handle;

pub fn handles(&self) -> Handles

Important traits for Collisions<'a>
impl<'a> Iterator for Collisions<'a> type Item = Collision<'a>;

pub fn collisions(&self) -> Collisions

pub fn get(&self, handle: Handle) -> Option<Collision>

pub fn begin_build(&mut self) -> CompoundBuilder