Trait newton::handle::IntoHandle

pub trait IntoHandle {
    fn into_handle(self, newton: &Newton) -> Handle;
}

Required methods

fn into_handle(self, newton: &Newton) -> Handle

Moves the object into the given Newton and returns a handle to borrow it or retake ownership of it later.

Loading content...