Struct newton::joint::UserJoint

pub struct UserJoint<'a> { /* fields omitted */ }

Generic user-defined joint (for advanced and very specific cases).

Methods

impl<'a> UserJoint<'a>

pub unsafe fn from_raw(raw: *const NewtonJoint, owned: bool) -> Self

pub fn release(self)

Forgets about this body. It will be freed along the world If you want to reference the object later, you should call into_handle instead.