Struct newton::Newton

pub struct Newton { /* fields omitted */ }

Newton dynamics context

Methods

impl Newton

pub unsafe fn from_raw(raw: *const NewtonWorld) -> Self

pub fn create() -> Self

pub fn set_solver(&mut self, steps: usize)

Sets the solver model.

steps = 0 Exact solver

step > 0 Linear solver with the given number of linear steps.

pub fn set_threads(&mut self, threads: usize)

Sets the number of threads Newton will run the simulation on. By default, the simulation is single-threaded.

pub fn threads(&self) -> usize

pub fn constraints(&self) -> usize

pub fn bodies(&self) -> usize

Important traits for Bodies<'a>
impl<'a> Iterator for Bodies<'a> type Item = Body<'a>;

pub fn bodies_iter(&self) -> Bodies

pub fn body(&self, handle: Handle) -> Option<Body>

pub fn body_take(&self, handle: Handle) -> Option<Body>

pub fn collision(&self, handle: Handle) -> Option<Collision>

pub fn collision_take(&self, handle: Handle) -> Option<Collision>

pub const fn as_raw(&self) -> *const NewtonWorld

pub fn into_raw(self) -> *const NewtonWorld

impl Newton

ffi wrappers

pub fn invalidate(&mut self)

Invalidated any cached contacts information.

Useful for when you synchronize a simulation over a network and need to run the simulation deterministically.

pub fn update(&mut self, step: Duration)

Steps the simulation by a fixed amount (synchronous).

pub fn update_async(&mut self, step: Duration) -> AsyncUpdate

Steps the simulation by a fixed amount (asynchronous) without blocking the current thread of execution.

Notes

If you intend to perform network synchronization, synchronous update is the preferred approach.

impl Newton

Ray & convex collision casting

pub fn ray_cast<F>(&self, p0: [f32; 3], p1: [f32; 3], filter: F, thread: usize) where
F: FnMut(Body, Collision, [f32; 3], [f32; 3], f32) -> f32,

Samples world with a ray, and runs the given filter closure for every body that intersects.

The filter closure can be used to implement various flavors of ray casting. Refer to the official wiki for further info.